Match a lot of `BrightObj` (#2000)
* Match a lot of `BrightObj`

* Updated from feedback

+++ include/Game/Screen/LensFlare.hpp
@@ -102,5 +102,5 @@ public:
 namespace MR {
     void addBrightObj(BrightObjBase* pBrightObj);
     void setLensFlareDrawSyncToken();
-    u16 getLensFlareDrawSyncTokenIndex();
+    u32 getLensFlareDrawSyncTokenIndex();
 };  // namespace MR
